feat: Evaluate kits on label text
Introducing the ability to invoke a kit function and use the
return value at a given position in a kitButton label.

To include the return value of a kit, wrap the kit name in pipes.
For example, "next appointment in |nextAppointment| days" returns
"next appointment in 3 days".

Can be combined with style expression filters.