feat(next appointment): Use activity icons
Show an activity icon before the event name.

The icon is retrieved from the event's linked
references in :block/properties { :activity }.

On the :activity page, the icon is drawn from
the :-icon page propertx. It must be a 4-character
tabler icon glyph code.

Bug: When multiple linked references are present
in the :activity property, ALL pages MUST posses
the :-icon page property. Otherwise, no icon is
displayed.