fix(advanced query): Discussion topics don't use aliases.
The dicussion topics queries does not return results when aliased
tags are used in conjunction with the primary page. Specifically,
the tag #topic has the alias #topics, but blocks taged with #topics
do not appear in the discussion list.

This change improved the advanced query to include pages aliased
to the primary topic tag.