fix(topic): Allow refs to #topic or an alias
Fixed query dealing with the tag-reference for topics.

The current query mistakenly required the #topic tag to have an
alias.
Instead, the new query uses an or-join to trigger topic-blocks
by including a linked reference to either the tag (#topic) or
an alias of the tag (e.g. #topics). AND, if the tag has no alias
it doesn't break the query.